refactor(console): share generated runtime writer
